requirements.txt
cffi==1.16.0
clr-loader==0.2.6
psutil==5.9.8
pycparser==2.21
pythonnet==3.0.3
